make FinallyDelegate more resilient regarding unhandled exceptions
The idea of `FinallyDelegate` is to capture exceptions that aren't
caught by the testcase. In order to realize that, it registers a handler
in the current appdomain. In case the handler is executed, it gets the
right context by relying on the fact that the testcase calls a `Set()`
method, which pushes the necessary context on a stack. After completion
of the testcase, `Complete()` is called and thus the context is removed
from the stack. (NB: TestResults can be nested, thus a stack
datastructure makes sense here).

Now consider the following scenario:

Testcase A is spawning another ThreadX, and ThreadX is throwing an
exception. The order of things can be like this:

* TestcaseA does `FinallyDelegate.Set()`
* TestcaseA executes
* ThreadX spawns
* TestcaseA finishes and calls `FinallyDelegate.Complete()`
* ThreadX throws exception and thus using the wrong context to report the error

In the last step three things can happen:

(1) Top of the stack is the "testsuite result" and records an error in the
    test suite. It is then really just marked as "failed", but it doesn't
    tell _what_ testcases failed. This ends up being confusing, as the test
    harness doesn't show any errors, but still returns with a non-zero exit
    code. For example that happend here
    https://github.com/mono/mono/pull/5911#issuecomment-341194418
    and here:
    https://jenkins.mono-project.com/job/test-mono-pull-request-arm64/5877/

(2) Top of the stack is another testcase, and consequently the error
    will be reported for the wrong testcase. Not nice.

(3) Stack is empty, and thus it would end in a NRE. I haven't seen this
    one yet.

That's about the state of affairs. How to solve this?

Thanks to @luhenry who pointed me to the right direction: We going to
use less-known feature called `ExecutionContext`. An `ExecutionContext`
is bound to managed thread. It has something called
`LogicalCallContext`, which has an interesting property: If a newly
Thread is created, it also gets a new `ExecutionContext` associated with
it, BUT ALSO inherits the `LogicalCallContext` from the calling thread.
We going to take advantage of this fact, and create explicitly a new
`ExecutionContext` for each test. That way, we can store the necessary
information in the `LogicalCallContext`, so that unhandled exceptions
can be assigned to the correct `TestResult` object by the
`FinallyDelegate`.